choer(): move proxy util out of hot path
