+++ .gitignore
@@ -10,3 +10,4 @@ testfile.ts
 privaterepotracker
 restClientRegex.ts
 .DS_Store
+dist
