+++ package.json
@@ -44,10 +44,10 @@
     "eslint": "^7.10.0",
     "jest": "^29.1.1",
     "ts-jest": "^29.1.1",
-    "typescript": "^4.7.4"
+    "typescript": "^4.7.4",
+    "webpack-cli": "^5.1.4"
   },
   "optionalDependencies": {
-    "https-proxy-agent": "^6.1.0",
     "source-map-loader": "^2.0.0",
     "ts-loader": "^8.0.11",
     "webpack": "^5.4.0"
