+++ src/index.ts
@@ -8,7 +8,6 @@ export * from './types/shared';
 export * from './types/spot';
 export * from './types/websockets';
 export * from './usdm-client';
-export * from './util/proxy';
 export * from './util/requestUtils';
 export * from './util/typeGuards';
 export * from './util/WsStore';
